// server/src/participant/enum/participant-error.enum.ts
export enum PARTICIPANT_ERROR {
  PROJECT_NOT_FOUND = "PARTICIPANT_ERROR_PROJECT_NOT_FOUND",
  ACCESS_DENIED = "PARTICIPANT_ERROR_ACCESS_FORBIDEN",
  ALREADY_REGISTERED = "PARTICIPANT_ERROR_ALREADY_REGISTERED",
  REGISTRATION_NOT_STARTED = "PHASE_ERROR_REGISTRATION_NOT_STARTED",
  REGISTRATION_ENDED = "PHASE_ERROR_REGISTRATION_ENDED",
  NO_REGISTRATION = "PHASE_ERROR_NO_REGISTRATION",
  NOT_ENOUGH_BALANCE = "PHASE_ERROR_NOT_ENOUGH_BALANCE",
  NOT_MEMBER_OF_REQUIRED_GUILD = "PARTICIPANT_ERROR_NOT_MEMBER_OF_REQUIRED_GUILD",
  HAS_NONE_OF_REQUIRED_ROLES = "PARTICIPANT_ERROR_HAS_NONE_OF_REQUIRED_ROLES",
  DISCORD_AUTH_TOKEN_EXPIRED = "PARTICIPANT_ERROR_DISCORD_AUTH_TOKEN_EXPIRED",
  TWITTER_API_MET_LIMIT = "PARTICIPANT_ERROR_TWITTER_API_MET_LIMIT",
  TWITTER_NOT_FOLLOWING = "PARTICIPANT_ERROR_TWITTER_NOT_FOLLOWING",
  DISCORD_NOT_CONNECTED = "PARTICIPANT_ERROR_DISCORD_NOT_CONNECTED",
  TWITTER_NOT_CONNECTED = "PARTICIPANT_ERROR_TWITTER_NOT_CONNECTED",
  EMAIL_REQUIRED = "PARTICIPANT_ERROR_EMAIL_REQUIRED",
  DISCORD_ERROR = "PARTICIPANT_ERROR_DISCORD_ERROR",
  PARTICIPANT_NOT_FOUND = "PARTICIPANT_ERROR_PARTICIPANT_NOT_FOUND",
  INVALID_PARTICIPANT_ADDRESS = "PARTICIPANT_ERROR_INVALID_PARTICIPANT_ADDRESS",
}